refactor: 응답 코드 리팩토링 및 파일명/ 경로 수정
