+++ src/controllers/studyController.js
@@ -1,17 +1,27 @@
 import prisma from '../lib/prisma.js';
 
 export const getStudies = async (req, res) => {
-  const studies = await prisma.study.findMany({
-    select: {
-      id: true,
-      nickname: true,
-      title: true,
-      description: true,
-      background: true,
-      createdAt: true,
-      updatedAt: true,
-    },
-  }); 
+  try {
+    const studies = await prisma.study.findMany({
+      select: {
+        id: true,
+        nickname: true,
+        title: true,
+        description: true,
+        background: true,
+        createdAt: true,
+        updatedAt: true,
+      },
+    });
 
-  res.json(studies);
+    res.status(200).json({
+      success: true,
+      data: studies,
+    });
+  } catch (error) {
+    res.status(400).json({
+      success: false,
+      message: error.message,
+    });
+  }
 };
